Incorporated some stuff from the Gatsby tutorials.
Added a Footer, a splash image for the home page, and populated the other pages on the website. I also moved the blog post listing onto the /blog/ page.

I am considering uninstalling emotion because I was having trouble dealing with LayoutMain and specialzing the  formatting of the home page. I like emotion, but I do not like these `StyledX` components. I might come back to it, but for now, I am going to prefer inline CSS.